Update code to that computes starting values
Moved to a MLE system for setting starting values to speed up the code.  We need to test to make sure that it works as well.  We are going to eventually want to replace all the bayesm dependencies.

I also put the code into the main algorithm so that users don’t have to manually compute starting values before runnign our code.